Rename files and helper function
